+++ packages/react/src/components/ToastContainer/styles.ts
@@ -1,0 +1,19 @@
+import * as ToastPrimitive from '@radix-ui/react-toast'
+
+import { styled } from '../../styles'
+
+export const ToastContainerViewport = styled(ToastPrimitive.Viewport, {
+  position: 'fixed',
+  bottom: 0,
+  right: 0,
+  display: 'flex',
+  flexDirection: 'column',
+  padding: '25px',
+  gap: '10px',
+  width: '390px',
+  maxWidth: '100vw',
+  margin: 0,
+  listStyle: 'none',
+  zIndex: 2147483647,
+  outline: 'none',
+})