+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @victordev-ignite-ui/react
 
+## 2.0.3
+
+### Patch Changes
+
+- new Toast component
+
+## 2.0.2
+
+### Patch Changes
+
+- new Tooltip component
+
 ## 2.0.1
 
 ### Patch Changes